Building Product class, updating "Get" function so it actually works... Required changing data types of "Unit_Cost" and "List_Price" for the time being as I could not convert from money to double

+++ Database/dbo/Tables/Product.sql
@@ -2,12 +2,14 @@
     [Product_Id]   INT          IDENTITY (1, 1) NOT NULL,
     [Product_Code] VARCHAR (20) NOT NULL,
     [Product_Name] NCHAR (10)   NOT NULL,
-    [Unit_Cost]    MONEY        NOT NULL,
-    [List_Price]   MONEY        NOT NULL,
+    [Unit_Cost]    FLOAT (53)   NOT NULL,
+    [List_Price]   FLOAT (53)   NOT NULL,
     [Discontinue]  BIT          DEFAULT ((0)) NOT NULL,
     [Category]     VARCHAR (50) NOT NULL,
     PRIMARY KEY CLUSTERED ([Product_Id] ASC)
 );
+
+
